+++ inst/extdata/run_seeker_array.R
@@ -1,6 +1,0 @@
-cArgs = commandArgs(TRUE)
-yamlPath = cArgs[1L]
-parentDir = cArgs[2L]
-
-params = yaml::read_yaml(yamlPath)
-seeker::seekerArray(params, parentDir)