+++ tests/testthat/test-array.R
@@ -6,7 +6,7 @@ withr::local_file(parentDirArr)
 test_that('checkSeekerArrayArgs', {
   skip_on_os('windows', arch = NULL)
 
-  resultObs = checkSeekerArrayArgs(paramsArray, parentDirArr)
+  resultObs = checkSeekerArrayArgs(paramsArray$study, paramsArray$geneIdType, paramsArray$platform, parentDirArr)
   resultExp = snapshot(
     resultObs, file.path(dataDir, 'seeker_array_args_output.qs'))
 
@@ -16,26 +16,20 @@ test_that('checkSeekerArrayArgs', {
 test_that('checkSeekerArrayArgs errors', {
   skip_on_os('windows', arch = NULL)
 
-  paramsArrayErr = paramsArray
-
   # GSE platform not null or GPL
-  paramsArrayErr$platform = 'abcd'
-  expect_error(checkSeekerArrayArgs(paramsArrayErr, parentDirArr))
+  expect_error(checkSeekerArrayArgs(paramsArray$study, paramsArray$geneIdType, 'abcd', parentDirArr))
 
   # E- with platform
-  paramsArrayErr$study = 'E-test'
-  expect_error(checkSeekerArrayArgs(paramsArrayErr, parentDirArr))
+  expect_error(checkSeekerArrayArgs('E-test', paramsArray$geneIdType, 'abcd', parentDirArr))
 
   # raw with raw dir not existing
-  paramsArrayErr$study = 'LOCAL'
-  paramsArrayErr$platform = 'GPL1'
-  expect_error(checkSeekerArrayArgs(paramsArrayErr, parentDirArr))
+  expect_error(checkSeekerArrayArgs('LOCAL', paramsArray$geneIdType, 'GPL1', parentDirArr))
 })
 
 test_that('seekerArray GSE', {
   skip_on_os('windows', arch = NULL)
 
-  seekerArray(paramsArray, parentDirArr)
+  seekerArray(params = paramsArray, parentDir = parentDirArr)
 
   resultObs = list.files(parentDirArr, recursive = TRUE)
   resultExp = snapshot(
@@ -53,7 +47,7 @@ test_that('seekerArray Ae', {
   paramsArrayAe = paramsArray
   paramsArrayAe$study = 'E-MTAB-8714'
 
-  seekerArray(paramsArrayAe, parentDirArrAe)
+  seekerArray('E-MTAB-8714', paramsArray$geneIdType, paramsArray$platform, parentDir = parentDirArrAe)
 
   resultObs = list.files(parentDirArrAe, recursive = TRUE)
   resultExp = snapshot(
@@ -71,7 +65,7 @@ test_that('seekerArray LOCAL', {
   paramsArrayLocal = yaml::read_yaml(file.path(dataDir, 'LOCAL01.yml'))
   file.copy(file.path(dataDir, 'LOCAL01'), parentDirArrLcl, recursive = TRUE)
 
-  seekerArray(paramsArrayLocal, parentDirArrLcl)
+  seekerArray(params = paramsArrayLocal, parentDir = parentDirArrLcl)
 
   resultObs = list.files(parentDirArrLcl, recursive = TRUE)
   resultExp = snapshot(